// 04-IMPL/mejoredu-portal/src/app/follow/spaa/adecuaciones/requests/request/arr/budgets-arr/interface/budget-calendar.interface.ts
export interface BudgetCalendarI {
    status: string;
    total: number;
    january: number;
    february: number;
    march: number;
    april: number;
    may: number;
    june: number;
    july: number;
    august: number;
    september: number;
    october: number;
    november: number;
    december: number;
}